requirements.txt
Django~=4.0.4
django-crispy-forms==1.9.2
Pillow==9.1.1
image==1.5.33